Spawn and game managers

+++ Assets/Scripts/Enemies/EnemyData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyData", menuName = "Scriptable Objects/EnemyData")]
 public class EnemyData : ScriptableObject
 {
-    public float health;
+    public int health;
     public float speed;
     public float damage;
     public float attackRange;
